Fix: 실제 사용자가 저장한 도서와 음악 개수, 작성한 서평과 음악 리뷰 개수가 반영되도록 수정

+++ roome/src/main/java/com/roome/domain/room/dto/UserStorageDto.java
@@ -9,12 +9,12 @@ import lombok.Getter;
 @AllArgsConstructor
 public class UserStorageDto {
 
-    private int savedMusic;
-    private int savedBooks;
-    private int writtenReviews;
-    private int writtenMusicLogs;
+    private Long savedMusic;
+    private Long savedBooks;
+    private Long writtenReviews;
+    private Long writtenMusicLogs;
 
-    public static UserStorageDto from(int savedMusic, int savedBooks, int writtenReviews, int writtenMusicLogs) {
+    public static UserStorageDto from(Long savedMusic, Long savedBooks, Long writtenReviews, Long writtenMusicLogs) {
         return UserStorageDto.builder()
                 .savedMusic(savedMusic)
                 .savedBooks(savedBooks)
